+++ compose/test/test_helper.rb
@@ -1,0 +1,9 @@
+$LOAD_PATH.unshift File.expand_path("../lib", __dir__)
+require "testcontainers/compose"
+
+require "minitest/autorun"
+require "minitest/hooks/test"
+
+class TestcontainersTest < Minitest::Test
+  include Minitest::Hooks
+end