+++ mariadb/lib/testcontainers/mariadb.rb
@@ -126,7 +126,7 @@ module Testcontainers
     end
 
     def _default_healthcheck_options
-      {test: ["/usr/bin/mysql", "--protocol=TCP", "--port=#{port}", "--user=#{username}", "--password=#{password}", database, "--silent", "--execute=SELECT 1;"], interval: 1, timeout: 5, retries: 5}
+      {test: ["/usr/local/bin/healthcheck.sh", "--connect"], interval: 1, timeout: 5, retries: 5}
     end
   end
 end