+++ mariadb/CHANGELOG.md
@@ -1,3 +1,11 @@
-## [Unreleased]
+## [0.1.1] - 2023-06-13
+
+### Fixed
+
+- Use healthcheck script shipped with the Docker image
+
+## [0.1.0] - 2023-06-10
 
 ### Added
+
+- Initial release of the MariaDB module