+++ mysql/CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.1.1] - 2023-06-10
+
+### Added
+
+- Add healthcheck support on start
+
+### Fixed
+
+- Remove support for custom ports, it is not supported by underlying
+  images.
+
 ## [0.1.0] - 2023-05-13
 
 ### Added