+++ redpanda/CHANGELOG.md
@@ -1,3 +1,12 @@
-## [Unreleased]
+## [0.2.0] - 2024-02-08
+
+### Fixed
+
+- Fixed startup script for the Redpanda module
+
+
+## [0.1.0] - 2023-06-10
 
 ### Added
+
+- Initial release of the Redpanda module